AtkRoot: stop to use G_DECLARE_DERIVABLE_TYPE
Although in theory new code should use this macro, AFAIU, it needs to
be used with code that is already using it, or at least we would need
to add on AtkObject some clear_ptr functions. For now it is easier to
define the type "in the old way".

TODO: investigate if there is a way to use the new macro.